Refactor preset strategy registry
